+++ packages/iov-bcp-types/types/connection.d.ts
@@ -24,7 +24,6 @@ export interface Account {
      * - Available after first transaction sent (e.g. Lisk, Tendermint, Ethereum)
      */
     readonly pubkey?: PublicKeyBundle;
-    readonly name?: string;
     readonly balance: ReadonlyArray<BcpCoin>;
 }
 export interface BcpTicker {