+++ packages/iov-bns/types/context.d.ts
@@ -13,7 +13,6 @@ export interface ChainData {
 /** Like BCP's Account but with no pubkey. Keep compatible to Account! */
 export interface WalletData {
     readonly address: Address;
-    readonly name?: string;
     readonly balance: ReadonlyArray<BcpCoin>;
 }
 export declare class Context {